Trait oak_functions_launcher::instance::LaunchedInstance

pub trait LaunchedInstance {
    fn wait<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn kill<'async_trait>(
        self: Box<Self>
    ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
    fn create_comms_channel<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Channel>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Defines the interface of a launched guest instance. Standardizes the interface of different implementations, e.g. a VM in which the guest is running or the guest running directly as a unix binary.

Required Methods

fn wait<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Wait for the guest instance process to finish.

fn kill<'async_trait>(
self: Box<Self>
) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>where
Self: 'async_trait,

Kill the guest instance.

fn create_comms_channel<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Box<dyn Channel>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Creates a channel to communicate with the guest instance.

Since different VMMs might use different comms channels, we leave it up to the VMM to create the channel rather than passing it in as part of the parameters.